Deprecate RPC getblocknumber
This RPC is exactly identical to getblockcount.  This duplication
dates back to commit dbf95a724a0376847d8e70b0295c2d9d60b01035 when
Satoshi created the RPC interface.

There's no need to have both, so we standardize on "count" which
matches the naming convention in getconnectioncount.

Following the tradition established with previously deprecated APIs,
getblocknumber continues to work, but it's not listed in the help
system.