test: Add a second Sapling note to WalletTests.ClearNoteWitnessCache
Quoting the documentation for `std::vector::operator[]`:

    Portable programs should never call this function with an argument
    n that is out of range, since this causes undefined behavior.

This test was doing just that: performing checks on a non-existent
second Sapling witness (duplicating the Sprout logic that checked two
notes, one of which was in the wallet). The test was instead reading
arbitrary memory after the witness that did exist; in most cases, this
memory was interpreted as a `boost::none` as expected, but in some cases
the memory was interpreted as a "real" witness.

Closes zcash/zcash#4445.

Co-authored-by: Ying Tong <yingtong@ethereum.org>