Categorize listaddresses result by source type.
Addresses managed by the zcashd wallet may be generated using
multiple different sources of randomness and/or the wallet's
HD seeds. Depending upon how addresses are generated, they may
divided into multiple sets, each associated with a separate
spend authority and treated as an independent pool of funds.
In the future, the root spend authority for a wallet will
be a mnemonic seed phrase; the API represented by this PR
anticipates that future and attempts to provide functionality
that will serve both current and future needs, as well as the
transition between them.